+++ changelog.md
@@ -20,6 +20,11 @@
 
 
 
+*	2020.11.1
+
+	>	remote-file.vue: convert CRLF to LF to avoid href errors on Windows.
+
+
 *	2020.10.30
 
 	+	staffSvg: added type of HashTable.